+++ tokio-threadpool/src/worker.rs
@@ -13,6 +13,8 @@ use worker_state::{
     WORKER_SIGNALED,
 };
 
+use tokio_executor;
+
 use std::cell::Cell;
 use std::marker::PhantomData;
 use std::rc::Rc;
@@ -20,8 +22,6 @@ use std::thread;
 use std::time::Instant;
 use std::sync::atomic::Ordering::{AcqRel, Acquire};
 use std::sync::Arc;
-
-use tokio_executor;
 
 /// Thread worker
 ///
@@ -365,9 +365,7 @@ impl Worker {
 
         // The first part of the sleep process is to transition the worker state
         // to "pushed". Now, it may be that the worker is already pushed on the
-        // sleeper stack, in which case, we don't push again. However, part of
-        // this process is also to do some final state checks to avoid entering
-        // the mutex if at all possible.
+        // sleeper stack, in which case, we don't push again.
 
         loop {
             let mut next = state;
@@ -376,6 +374,9 @@ impl Worker {
                 WORKER_RUNNING => {
                     // Try setting the pushed state
                     next.set_pushed();
+
+                    // Transition the worker state to sleeping
+                    next.set_lifecycle(WORKER_SLEEPING);
                 }
                 WORKER_NOTIFIED | WORKER_SIGNALED => {
                     // No need to sleep, transition back to running and move on.
@@ -417,66 +418,18 @@ impl Worker {
             state = actual;
         }
 
-        // Acquire the sleep mutex, the state is transitioned to sleeping within
-        // the mutex in order to avoid losing wakeup notifications.
-        let mut lock = self.entry().park_mutex.lock().unwrap();
-
-        // Transition the state to sleeping, a CAS is still needed as other
-        // state transitions could happen unrelated to the sleep / wakeup
-        // process. We also have to redo the lifecycle check done above as
-        // the state could have been transitioned before entering the mutex.
-        loop {
-            let mut next = state;
-
-            match state.lifecycle() {
-                WORKER_RUNNING => {}
-                WORKER_NOTIFIED | WORKER_SIGNALED => {
-                    // Release the lock, sleep will not happen this call.
-                    drop(lock);
-
-                    // Transition back to running
-                    loop {
-                        let mut next = state;
-                        next.set_lifecycle(WORKER_RUNNING);
-
-                        let actual = self.entry().state.compare_and_swap(
-                            state.into(), next.into(), AcqRel).into();
-
-                        if actual == state {
-                            return true;
-                        }
-
-                        state = actual;
-                    }
-                }
-                _ => unreachable!(),
-            }
-
-            trace!(" sleeping -- set WORKER_SLEEPING; idx={}", self.idx);
-
-            next.set_lifecycle(WORKER_SLEEPING);
-
-            let actual = self.entry().state.compare_and_swap(
-                state.into(), next.into(), AcqRel).into();
-
-            if actual == state {
-                break;
-            }
-
-            state = actual;
-        }
-
         trace!("    -> starting to sleep; idx={}", self.idx);
 
         let sleep_until = self.inner.config.keep_alive
             .map(|dur| Instant::now() + dur);
 
-        // The state has been transitioned to sleeping, we can now wait on the
-        // condvar. This is done in a loop as condvars can wakeup spuriously.
+        // The state has been transitioned to sleeping, we can now wait by
+        // calling the parker. This is done in a loop as condvars can wakeup
+        // spuriously.
         loop {
             let mut drop_thread = false;
 
-            lock = match sleep_until {
+            match sleep_until {
                 Some(when) => {
                     let now = Instant::now();
 
@@ -486,14 +439,20 @@ impl Worker {
 
                     let dur = when - now;
 
-                    self.entry().park_condvar
-                        .wait_timeout(lock, dur)
-                        .unwrap().0
+                    unsafe {
+                        (*self.entry().park.get())
+                            .park_timeout(dur)
+                            .unwrap();
+                    }
                 }
                 None => {
-                    self.entry().park_condvar.wait(lock).unwrap()
+                    unsafe {
+                        (*self.entry().park.get())
+                            .park()
+                            .unwrap();
+                    }
                 }
-            };
+            }
 
             trace!("    -> wakeup; idx={}", self.idx);
 
@@ -504,9 +463,6 @@ impl Worker {
                 match state.lifecycle() {
                     WORKER_SLEEPING => {}
                     WORKER_NOTIFIED | WORKER_SIGNALED => {
-                        // Release the lock, done sleeping
-                        drop(lock);
-
                         // Transition back to running
                         loop {
                             let mut next = state;
@@ -526,6 +482,7 @@ impl Worker {
                 }
 
                 if !drop_thread {
+                    // This goees back to the outer loop.
                     break;
                 }