+++ tokio-executor/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.2 (unreleased)
+
+* Implement `Unpark` for `Box<Unpark>`.
+
 # 0.1.1 (March 22, 2018)
 
 * Optionally support futures 0.2.